fix: penalty coefficient option changed form int to float

+++ utils/options.py
@@ -138,8 +138,8 @@ class Options:
         self.parser.add_argument('--soft_tau', type=float, default=0.005, help='Tau for soft update')
 
         self.parser.add_argument('--mu_penalty', type=int, default=1, help='Penalty applied to mu during policy optimisation -> 1 for penalty applied, 0 for no penalty')
-        self.parser.add_argument('--action_penalty_limit', type=int, default=1, help='Limit of action during policy optimisation')
-        self.parser.add_argument('--action_penalty_coef', type=int, default=0.1, help='Policy penalty term coefficient')
+        self.parser.add_argument('--action_penalty_limit', type=float, default=1, help='Limit of action during policy optimisation')
+        self.parser.add_argument('--action_penalty_coef', type=float, default=0.1, help='Policy penalty term coefficient')
 
         self.parser.add_argument('--replay_buffer_type', type=str, default="random", help='type of replay buffer')
         self.parser.add_argument('--replay_buffer_alpha', type=float, default=0.6, help='Replay buffer alpha')
